make 'pkgs' error non-fatal; better error handling
tllocalmgr uses 'pkgs' files in order to get version information about
what texlive/CTAN packages are installed on the system. For official
texlive-* packages and some other ones this works fine, but a lot of
AUR packages *do not* create the a 'pkgs' file. In this situation,
tllocalmgr would fatal error out.

This is a little extreme, though sensible as without a valid 'pkgs' file
to parse for LaTeX package versions, tllocalmgr _might_ allow for
packages to be installed which are already present on the machine. This
however is _unlikly_ as when installing through pacman, we should hit
into the FS integrity check, causing install to fail.

Ideally we would want all texlive packages to always have a 'pkgs' file
but that is unlikely. An alternative idea might be to create a
pacman-hook to generate one if missing, but could be very tricky and
time consuming.

* we now ignore any missing pkgs file, instead issueing a warning to the
  user to check upstream to have this fixed
* we also properly check that a package actually exists in the tlpdb
  *before* trying to read its metadata.